+++ src/biowdl_input_converter/__init__.py
@@ -27,6 +27,8 @@ import argparse
 from pathlib import Path
 
 from . import input_conversions, output_conversions
+from .utils import check_duplicate_files, check_existence_list_of_files, \
+    check_md5sums
 
 
 def argument_parser() -> argparse.ArgumentParser:
@@ -34,7 +36,11 @@ def argument_parser() -> argparse.ArgumentParser:
         description="Parse samplesheets for BioWDL pipelines.")
     parser.add_argument("samplesheet", type=str,
                         help="The input samplesheet. Format will be "
-                             "automatically detected.")
+                             "automatically detected from file suffix "
+                             "if --format argument not provided")
+    parser.add_argument("-f", "--format", type=str,
+                        help="The input samplesheet format - "
+                              "tsv, csv, json, or yaml")
     parser.add_argument("-o", "--output",
                         help="The output file to which the json is written. "
                              "Default: stdout")
@@ -44,9 +50,14 @@ def argument_parser() -> argparse.ArgumentParser:
     parser.add_argument("--old", action="store_true", dest="old_style_json",
                         help="Output old style JSON as used in BioWDL "
                              "germline-DNA and RNA-seq version 1 pipelines")
-    parser.add_argument("--skip-file-check", action="store_true",
+    parser.add_argument("--skip-file-check", action="store_false",
+                        dest="file_check",
                         help="Skip the checking if files in the samplesheet "
                              "are present.")
+    parser.add_argument("--skip-duplicate-check", action="store_false",
+                        dest="duplicate_check",
+                        help="Skip the checks for duplicate files in the "
+                             "samplesheet.")
     parser.add_argument("--check-file-md5sums", action="store_true",
                         help="Do a md5sum check for reads which have md5sums "
                              "added in the samplesheet.")
@@ -54,23 +65,31 @@ def argument_parser() -> argparse.ArgumentParser:
 
 
 def samplesheet_to_json(samplesheet: Path,
+                        fileformat: str = None,
                         old_style_json: bool = False,
                         file_presence_check: bool = True,
-                        file_md5_check: bool = False) -> str:
+                        file_md5_check: bool = False,
+                        file_duplication_check: bool = True) -> str:
     """
     Converts a samplesheet file to JSON
     :param samplesheet:
+    :param fileformat: tsv, csv, yaml, yml, json
     :param old_style_json: Return a BioWDL old-style pipeline JSON
     :param file_presence_check: Check if the files in the samplesheet are
     present
     :param file_md5_check: Check if the md5sums for the files are correct
     :return: a JSON string presenting the BioWDL JSON.
     """
-    if samplesheet.suffix in [".tsv", ".csv"]:
+    if fileformat is not None:
+        filetype = fileformat.lower().replace('.', '')
+    else:
+        filetype = samplesheet.suffix.lower().replace('.', '')
+
+    if filetype in ["tsv", "csv"]:
         samplegroup = input_conversions.samplesheet_csv_to_samplegroup(
             samplesheet)
     # JSON can also be parsed by a YAML parser.
-    elif samplesheet.suffix in [".yaml", ".yml", ".json"]:
+    elif filetype in ["yaml", "yml", "json"]:
         samplegroup = input_conversions.biowdl_yaml_to_samplegroup(
             samplesheet)
     else:
@@ -78,9 +97,14 @@ def samplesheet_to_json(samplesheet: Path,
             f"Unsupported extension: {samplesheet.suffix}")
 
     if file_presence_check:
-        samplegroup.test_files_exist()
+        check_existence_list_of_files(samplegroup.files())
     if file_md5_check:
-        samplegroup.test_file_checksums()
+        files_with_sums = ((file, sum) for file, sum in
+                           samplegroup.files_and_md5sums()
+                           if sum is not None)
+        check_md5sums(files_with_sums)
+    if file_duplication_check:
+        check_duplicate_files(samplegroup.files())
 
     if old_style_json:
         output_json = output_conversions.samplegroup_to_biowdl_old_json(
@@ -93,10 +117,12 @@ def samplesheet_to_json(samplesheet: Path,
 
 def main():
     args = argument_parser().parse_args()
+
     output_json = samplesheet_to_json(
         samplesheet=Path(args.samplesheet),
         old_style_json=args.old_style_json,
-        file_presence_check=not args.skip_file_check,
+        file_presence_check=args.file_check,
+        file_duplication_check=args.duplicate_check,
         file_md5_check=args.check_file_md5sums)
 
     # Only generate output if not validating.